ref: rename neovim to nvim
