+++ client/src/nvim/state.rs
